Time: 434 ms (50.00%), Space: 43.5 MB (100.00%) - LeetHub

+++ 2328-number-of-increasing-paths-in-a-grid/2328-number-of-increasing-paths-in-a-grid.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    long long mod = 1e9 + 7;
+    int mod = 1e9 + 7;
     
     long long f(int i,int j,int prev,vector<vector<int>> &grid,vector<vector<int>> &dp){
         if(i< 0 or j< 0 or i >= grid.size() or  j >= grid[i].size() or grid[i][j] <= prev){
